+++ packages/protocol/test/layer2/TaikoL2Test.sol
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.24;
+
+import "../../contracts/layer2/DelegateOwner.sol";
+import "../layer2/TaikoL2EIP1559Configurable.sol";
+import "../layer2/LibL2Signer.sol";
+import "../shared/TaikoTest.sol";
+
+abstract contract TaikoL2Test is TaikoTest { }