+++ packages/taiko-client/bindings/encoding/protocol_config.go
@@ -20,7 +20,7 @@ var (
 		StateRootSyncInternal: 16,
 		MaxAnchorHeightOffset: 64,
 		OntakeForkHeight:      2,
-		BaseFeeConfig: bindings.TaikoDataBaseFeeConfig{
+		BaseFeeConfig: bindings.LibSharedDataBaseFeeConfig{
 			AdjustmentQuotient:     8,
 			SharingPctg:            75,
 			GasIssuancePerSecond:   5_000_000,
@@ -38,7 +38,7 @@ var (
 		StateRootSyncInternal: 16,
 		MaxAnchorHeightOffset: 64,
 		OntakeForkHeight:      840_512,
-		BaseFeeConfig: bindings.TaikoDataBaseFeeConfig{
+		BaseFeeConfig: bindings.LibSharedDataBaseFeeConfig{
 			AdjustmentQuotient:     8,
 			SharingPctg:            75,
 			GasIssuancePerSecond:   5_000_000,
@@ -57,7 +57,7 @@ var (
 		MaxAnchorHeightOffset: 64,
 		// TODO: update this value when mainnet fork height is decided
 		OntakeForkHeight: 9_000_000,
-		BaseFeeConfig: bindings.TaikoDataBaseFeeConfig{
+		BaseFeeConfig: bindings.LibSharedDataBaseFeeConfig{
 			AdjustmentQuotient:     8,
 			GasIssuancePerSecond:   5_000_000,
 			MinGasExcess:           1_340_000_000,