+++ packages/taiko-client/bindings/metadata/metadata_legacy.go
@@ -133,8 +133,8 @@ func (m *TaikoDataBlockMetadataLegacy) GetBlobIndex() uint8 {
 }
 
 // GetBaseFeeConfig returns the L2 block basefee configs.
-func (m *TaikoDataBlockMetadataLegacy) GetBaseFeeConfig() *bindings.TaikoDataBaseFeeConfig {
-	return &bindings.TaikoDataBaseFeeConfig{}
+func (m *TaikoDataBlockMetadataLegacy) GetBaseFeeConfig() *bindings.LibSharedDataBaseFeeConfig {
+	return &bindings.LibSharedDataBaseFeeConfig{}
 }
 
 // GetRawBlockHeight returns the raw L1 block height.