+++ packages/taiko-client/pkg/rpc/methods.go
@@ -273,7 +273,8 @@ func (c *Client) CalculateBaseFee(
 	l2Head *types.Header,
 	anchorBlockID *big.Int,
 	isOntake bool,
-	baseFeeConfig *bindings.TaikoDataBaseFeeConfig,
+	baseFeeConfig *bindings.LibSharedDataBaseFeeConfig,
+	currentTimestamp uint64,
 ) (*big.Int, error) {
 	var (
 		baseFeeInfo struct {
@@ -320,9 +321,9 @@ func (c *Client) CalculateBaseFee(
 		baseFeeInfo, err = c.TaikoL2.CalculateBaseFee(
 			&bind.CallOpts{BlockNumber: l2Head.Number, Context: ctx},
 			*baseFeeConfig,
-			uint64(time.Now().Unix())-l2Head.Time,
+			currentTimestamp-l2Head.Time,
 			parentGasExcess,
-			uint32(l2Head.GasUsed), // #nosec G115
+			uint32(l2Head.GasUsed),
 		)
 		if err != nil {
 			return nil, err
@@ -331,7 +332,7 @@ func (c *Client) CalculateBaseFee(
 		baseFeeInfo, err = c.TaikoL2.GetBasefee(
 			&bind.CallOpts{BlockNumber: l2Head.Number, Context: ctx},
 			anchorBlockID.Uint64(),
-			uint32(l2Head.GasUsed), // #nosec G115
+			uint32(l2Head.GasUsed),
 		)
 		if err != nil {
 			return nil, err
@@ -384,6 +385,7 @@ func (c *Client) GetPoolContent(
 		l1Head.Number,
 		chainConfig.IsOntake(l2Head.Number),
 		&chainConfig.ProtocolConfigs.BaseFeeConfig,
+		uint64(time.Now().Unix()),
 	)
 	if err != nil {
 		return nil, err
@@ -691,7 +693,6 @@ func (c *Client) checkSyncedL1SnippetFromAnchor(
 		return true, nil
 	}
 
-	// #nosec G115
 	if parentGasUsed != uint32(parent.GasUsed()) {
 		log.Info(
 			"Reorg detected due to parent gas used mismatch",